Revert ".storiesファイル移動してまとめた。"
This reverts commit 62e73101d7a37f19c2673f34c894235b0d2b1110.
