+++ src/components/atoms/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { Button } from "../Button";
+import { Button } from "./Button";
 
 const meta = {
     title: "atoms/Button",
