+++ src/components/atoms/symbols/CloseIcon.stories.tsx
@@ -1,19 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { SearchIcon } from "../symbols/SearchIcon";
+import { CloseIcon } from "./CloseIcon";
 
 const meta = {
-    title: "atoms/symbol/SearchIcon",
-    component: SearchIcon,
+    title: "atoms/symbol/CloseIcon",
+    component: CloseIcon,
     parameters: {
         layout: "centered",
     },
     tags: ["autodocs"],
     argTypes: {},
-} satisfies Meta<typeof SearchIcon>;
+} satisfies Meta<typeof CloseIcon>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const Search: Story = {}
+export const Cross: Story = {
+    args: {}
+}
