+++ src/components/atoms/symbols/QTextLogo.stories.tsx
@@ -1,21 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { QImageLogo } from "../symbols/QImageLogo";
+import { QTextLogo } from "./QTextLogo";
 
 const meta = {
-    title: "atoms/symbol/QImageLogo",
-    component: QImageLogo,
+    title: "atoms/symbol/QTextLogo",
+    component: QTextLogo,
     parameters: {
         layout: "centered",
     },
     tags: ["autodocs"],
     argTypes: {},
-} satisfies Meta<typeof QImageLogo>;
+} satisfies Meta<typeof QTextLogo>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
 
-export const QLogo: Story = {
+export const ThumbnailPicture: Story = {
     args: {},
-}
+};
