+++ src/components/molecules/SearchBar.stories.tsx
@@ -1,19 +1,18 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { TagIcon } from "../symbols/TagIcon";
+import { SearchBar } from "./SearchBar";
 
 const meta = {
-    title: "atoms/symbol/TagIcon",
-    component: TagIcon,
+    title: "SearchBar",
+    component: SearchBar,
     parameters: {
         layout: "centered",
     },
     tags: ["autodocs"],
-    argTypes: {},
-} satisfies Meta<typeof TagIcon>;
+    argTypes: {
+    },
+} satisfies Meta<typeof SearchBar>;
 
 export default meta;
 
 type Story = StoryObj<typeof meta>;
-
-export const Tag: Story = {};
