--- src/components/organisms/stories/ArticleCard.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-
-import { ArticleCard } from "../ArticleCard";
-
-const meta = {
-    title: "SearchBar",
-    component: ArticleCard,
-    parameters: {
-        layout: "centered",
-    },
-    tags: ["autodocs"],
-    argTypes: {},
-} satisfies Meta<typeof ArticleCard>;
-
-export default meta;
-
-type Story = StoryObj<typeof meta>;
